Wip : Refactor Details and Map
