+++ src/app/features/scpi/details-detials/details-detials.component.ts
@@ -1,15 +1,15 @@
-import { Component, inject, Input, OnInit } from '@angular/core';
-import { IftaLabelModule } from 'primeng/iftalabel';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { ButtonModule } from 'primeng/button';
-import { CardModule } from 'primeng/card';
-import { DividerModule } from 'primeng/divider';
-import { PanelModule } from 'primeng/panel';
-import { Details } from '../../../core/model/Details';
-import { DetailsDetailsService } from '../../../core/service/details-details.service';
-import { StatYear } from '../../../core/model/StatYear';
-import { Sector } from '../../../core/model/Sector';
+import {Component, inject, Input, OnInit} from '@angular/core';
+import {IftaLabelModule} from 'primeng/iftalabel';
+import {CommonModule} from '@angular/common';
+import {FormsModule} from '@angular/forms';
+import {ButtonModule} from 'primeng/button';
+import {CardModule} from 'primeng/card';
+import {DividerModule} from 'primeng/divider';
+import {PanelModule} from 'primeng/panel';
+import {Details} from '../../../core/model/Details';
+import {DetailsDetailsService} from '../../../core/service/details-details.service';
+import {StatYear} from '../../../core/model/StatYear';
+import {Sector} from '../../../core/model/Sector';
 
 @Component({
   selector: 'app-details-detials',
@@ -24,16 +24,17 @@ export class DetailsDetialsComponent implements OnInit {
   detailsService = inject(DetailsDetailsService);
   prix_retrait: number = 0;
   @Input() id_parent: number = -2000;
+
   ngOnInit(): void {
     this.getTheDetails(this.id_parent);
   }
+
   getTheDetails(id: number) {
     this.detailsService.getDetailsScpi(this.id_parent).subscribe({
       next: (res) => {
         if (res) {
           this.details = res;
-          this.stat = this.details.statYears && this.details.statYears.length > 0
-            ? this.details.statYears.reduce((prev, current) => (prev.yearStat.yearStat > current.yearStat.yearStat ? prev : current)) : null
+          this.stat = this.detailsService.getLastStats(this.details);
           if (this.stat) {
             this.prix_retrait = this.stat?.sharePrice - (this.stat?.sharePrice * (1 - this.details.subscriptionFees))
           } else {
