+++ src/app/features/scpi/details-global-view/details-global-view.component.ts
@@ -1,24 +1,17 @@
-import { CommonModule, isPlatformBrowser } from '@angular/common';
-import {
-  ChangeDetectorRef,
-  Component,
-  inject,
-  Input,
-  OnInit,
-  PLATFORM_ID,
-} from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { ButtonModule } from 'primeng/button';
-import { CardModule } from 'primeng/card';
-import { ChartModule } from 'primeng/chart';
-import { DividerModule } from 'primeng/divider';
-import { PanelModule } from 'primeng/panel';
-import { Details } from '../../../core/model/Details';
-import { Location } from '../../../core/model/Location';
-import { Sector } from '../../../core/model/Sector';
-import { StatYear } from '../../../core/model/StatYear';
-import { DetailsDetailsService } from '../../../core/service/details-details.service';
-import { MapComponent } from './map/map.component';
+import {CommonModule, isPlatformBrowser} from '@angular/common';
+import {ChangeDetectorRef, Component, inject, Input, OnInit, PLATFORM_ID,} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {ButtonModule} from 'primeng/button';
+import {CardModule} from 'primeng/card';
+import {ChartModule} from 'primeng/chart';
+import {DividerModule} from 'primeng/divider';
+import {PanelModule} from 'primeng/panel';
+import {Details} from '../../../core/model/Details';
+import {Location} from '../../../core/model/Location';
+import {Sector} from '../../../core/model/Sector';
+import {StatYear} from '../../../core/model/StatYear';
+import {DetailsDetailsService} from '../../../core/service/details-details.service';
+import {MapComponent} from '@/shared/component/map/map.component';
 
 @Component({
   selector: 'app-details-global-view',
@@ -48,7 +41,9 @@ export class DetailsGlobalViewComponent implements OnInit {
   platformId = inject(PLATFORM_ID);
 
   detailsService = inject(DetailsDetailsService);
-  constructor(private cd: ChangeDetectorRef) { }
+
+  constructor(private cd: ChangeDetectorRef) {
+  }
 
   ngOnInit(): void {
     this.getScpiGlobalInformation();
@@ -58,13 +53,7 @@ export class DetailsGlobalViewComponent implements OnInit {
     this.detailsService.getDetailsScpi(this.id_parent).subscribe(
       (res) => {
         this.details = res;
-        this.states =
-          this.details.statYears && this.details.statYears.length > 0
-            ? this.details.statYears.reduce((prev, current) =>
-              prev.yearStat.yearStat > current.yearStat.yearStat ? prev : current
-            )
-            : null;
-
+        this.states = this.detailsService.getLastStats(this.details);
         this.ListeLocations = this.details.locations;
         this.ListeSectors = this.details.sectors;
         this.initChart(this.ListeLocations);
@@ -171,6 +160,6 @@ export class DetailsGlobalViewComponent implements OnInit {
       return (value / 1_000).toFixed(1) + 'K';
     }
     return value.toString();
- }
+  }
 
 }
