+++ src/app/features/scpi/details/details.component.html
@@ -6,9 +6,9 @@
     <div>
       <div>
         <div class="card-container">
-          <p class="name_scpi">{{ details?.name }}</p>
+          <p class="name_scpi">{{ details?.name || "N/A" }}</p>
           <div class="p_texte">
-            <p class="small-label">Bureau</p>
+            <p class="small-label">{{mostSector?.id?.name || "N/A"}}</p>
             <p class="small-label">Europe</p>
           </div>
           <div class="p-grid">
