+++ src/app/features/scpi/details/details.component.ts
@@ -1,19 +1,21 @@
-import { CommonModule, Location } from '@angular/common';
-import { Component, inject, OnInit} from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { ButtonModule } from 'primeng/button';
-import { CardModule } from 'primeng/card';
-import { Details } from '../../../core/model/Details';
-import { StatYear } from '../../../core/model/StatYear';
-import { DetailsDetailsService } from '../../../core/service/details-details.service';
-import { DetailsDetialsComponent } from '../details-detials/details-detials.component';
-import { DetailsGlobalViewComponent } from '../details-global-view/details-global-view.component';
-import { ActivatedRoute } from '@angular/router';
-import { TabViewModule } from 'primeng/tabview';
-import { TranslateModule } from '@ngx-translate/core';
-import { ScpiHistoryDetailsComponent } from '../details-history/scpi-history-details.component';
-import { MessageService } from 'primeng/api';
-import { ToastModule } from 'primeng/toast';
+import {CommonModule, Location} from '@angular/common';
+import {Component, inject, OnInit} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {ButtonModule} from 'primeng/button';
+import {CardModule} from 'primeng/card';
+import {Details} from '../../../core/model/Details';
+import {StatYear} from '../../../core/model/StatYear';
+import {DetailsDetailsService} from '../../../core/service/details-details.service';
+import {DetailsDetialsComponent} from '../details-detials/details-detials.component';
+import {DetailsGlobalViewComponent} from '../details-global-view/details-global-view.component';
+import {ActivatedRoute} from '@angular/router';
+import {TabViewModule} from 'primeng/tabview';
+import {TranslateModule} from '@ngx-translate/core';
+import {ScpiHistoryDetailsComponent} from '../details-history/scpi-history-details.component';
+import {MessageService} from 'primeng/api';
+import {ToastModule} from 'primeng/toast';
+import {Sector} from '@/core/model/Sector';
+
 @Component({
   selector: 'app-details',
   imports: [
@@ -40,12 +42,11 @@ export class DetailsComponent implements OnInit {
   stat: StatYear | null = null;
   id_parent: number = -1;
   id_string: string | null = '';
-  retour: string = 'Details.RETOUR';
   vue: string = 'Details.VUE';
   history: string = 'Details.HISTORIQUE';
   _details: string = 'Details.DETAILS';
 
-
+  mostSector: Sector | null = null;
   activeTab: string = 'vue';
 
   setActiveTab(tab: string) {
@@ -56,7 +57,8 @@ export class DetailsComponent implements OnInit {
     private _location: Location,
     private route: ActivatedRoute,
     private messageService: MessageService
-  ) {}
+  ) {
+  }
 
   ngOnInit(): void {
     this.id_string = this.route.snapshot.paramMap.get('id');
@@ -67,16 +69,10 @@ export class DetailsComponent implements OnInit {
 
   loadNameRandMin() {
     this.detailsService.getDetailsScpi(this.id_parent).subscribe(
-      (res) => {
+      (res: Details) => {
         this.details = res;
-        this.stat =
-          this.details.statYears && this.details.statYears.length > 0
-            ? this.details.statYears.reduce((prev, current) =>
-                prev.yearStat.yearStat > current.yearStat.yearStat
-                  ? prev
-                  : current
-              )
-            : null;
+        this.stat = this.detailsService.getLastStats(this.details);
+        this.mostSector = this.detailsService.getMaxSectorLastYear(this.details);
       },
       () => {
         this.messageService.add({
