+++ src/app/shared/component/map/map.component.spec.ts
